:truck: Move ProcessorStage logic into EncodeProcessDecode. This allows finetuning on processors with a latent-space loss to dispatch correctly during training.
