+++ icenet_mp/models/base_model.py
@@ -242,9 +242,7 @@ class BaseModel(LightningModule, ABC):
         batch = self.process_batch(batch)
         target = batch["target"].clone().detach()
         prediction = self(batch)
-        loss = getattr(self, "_last_processor_loss", None)
-        if loss is None:
-            loss = self.loss(prediction, target)
+        loss = self.loss(prediction, target)
 
         # Log metrics; computation will be done at epoch end
         self.log(
